EMS-644 moved mapping to another function and fixed failing test
