+++ src/ui/server/helpers/mappings/map-financial-year-end-date.ts
@@ -1,6 +1,6 @@
-import formatDate from '../../../../../helpers/date/format-date';
-import { FIELD_IDS } from '../../../../../constants';
-import { FIELDS } from '../../../../../content-strings/fields/insurance/your-business';
+import formatDate from '../date/format-date';
+import { FIELD_IDS } from '../../constants';
+import { FIELDS } from '../../content-strings/fields/insurance/your-business';
 
 const { FINANCIAL_YEAR_END_DATE } = FIELD_IDS.INSURANCE.EXPORTER_BUSINESS.TURNOVER;
